requirements.txt
line-bot-sdk
google-generativeai
Flask
